# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceCaptureBody": {
        "description": "Request body for invoice capture.",
        "properties": {
          "cfdi_use": {
            "default": "G03",
            "description": "Uso del CFDI: G03 (servicios profesionales), P01 (honorarios)",
            "maxLength": 4,
            "minLength": 3,
            "title": "Cfdi Use",
            "type": "string"
          },
          "rfc": {
            "description": "RFC del cliente (receptor). Ej: XAXX010101000",
            "maxLength": 13,
            "minLength": 12,
            "title": "Rfc",
            "type": "string"
          }
        },
        "required": [
          "rfc"
        ],
        "title": "InvoiceCaptureBody",
        "type": "object"
      },
      "InvoiceRead": {
        "description": "Response shape for an invoice.",
        "properties": {
          "cfdi_use": {
            "title": "Cfdi Use",
            "type": "string"
          },
          "cfdi_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cfdi Uuid"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "iva": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Iva",
            "type": "string"
          },
          "payment_id": {
            "title": "Payment Id",
            "type": "string"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "quotation_id": {
            "title": "Quotation Id",
            "type": "string"
          },
          "rfc_emisor": {
            "title": "Rfc Emisor",
            "type": "string"
          },
          "rfc_receptor": {
            "title": "Rfc Receptor",
            "type": "string"
          },
          "stamped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stamped At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal",
            "type": "string"
          },
          "total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "type": "string"
          },
          "xml_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xml Url"
          }
        },
        "required": [
          "id",
          "payment_id",
          "quotation_id",
          "rfc_emisor",
          "rfc_receptor",
          "cfdi_use",
          "subtotal",
          "iva",
          "total",
          "status",
          "cfdi_uuid",
          "pdf_url",
          "xml_url",
          "stamped_at",
          "error_message",
          "created_at"
        ],
        "title": "InvoiceRead",
        "type": "object"
      },
...
      "PaymentCreateDeposit": {
        "additionalProperties": false,
        "description": "Admin records a deposit payment.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "id_auction": {
            "title": "Id Auction",
            "type": "string"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "required": [
          "id_auction",
          "amount"
        ],
        "title": "PaymentCreateDeposit",
        "type": "object"
      },
      "PaymentCreateStripe": {
        "additionalProperties": false,
        "description": "Admin creates a Stripe payment preference for a quotation.",
        "properties": {
          "id_auction": {
            "title": "Id Auction",
            "type": "string"
          }
        },
        "required": [
          "id_auction"
        ],
        "title": "PaymentCreateStripe",
        "type": "object"
      },
      "PaymentRead": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "auction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auction Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "quotation_id": {
            "title": "Quotation Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "stripe_checkout_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Checkout Session Id"
          },
          "stripe_payment_intent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Payment Intent Id"
          },
          "stripe_payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Payment Status"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "quotation_id",
          "type",
          "state",
          "currency",
          "created_at",
          "updated_at"
        ],
        "title": "PaymentRead",
        "type": "object"
      },
      "PaymentRefundBody": {
        "description": "Request body for refunding a payment.",
        "properties": {
          "amount_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Cents"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "PaymentRefundBody",
        "type": "object"
      },
      "PaymentRefundResponse": {
        "description": "Response after a refund is issued.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "message"
        ],
        "title": "PaymentRefundResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FastAPI port of the Mobbit B2B/B2C marketplace Rust Lambdas. See `docs/research/business-domain.md` for the business context and `docs/research/deployment-drift.md` for the original infra.",
    "title": "Mobbit Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/invoice": {
      "get": {
        "operationId": "list_invoices_api_admin_invoice_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceRead"
                  },
                  "title": "Response List Invoices Api Admin Invoice Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List all invoices (admin)",
        "tags": [
          "admin:invoices"
        ]
      }
    },
    "/api/admin/payment": {
      "get": {
        "operationId": "list_all_payments_api_admin_payment_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaymentRead"
                  },
                  "title": "Response List All Payments Api Admin Payment Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List all payments",
        "tags": [
          "admin:payments"
        ]
      }
    },
    "/api/admin/payment/{payment_id}": {
      "get": {
        "operationId": "get_payment_api_admin_payment__payment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a single payment",
        "tags": [
          "admin:payments"
        ]
      }
    },
    "/api/admin/payment/{payment_id}/refund": {
      "post": {
        "description": "Issue a full or partial refund for a PAID payment.",
        "operationId": "refund_payment_api_admin_payment__payment_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRefundBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRefundResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refund a payment (admin)",
        "tags": [
          "admin:payments"
        ]
      }
    },
...
    "/api/admin/quotation/{quotation_id}/payment/deposito": {
      "post": {
        "operationId": "create_deposit_payment_api_admin_quotation__quotation_id__payment_deposito_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreateDeposit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record a deposit payment",
        "tags": [
          "admin:payments"
        ]
      }
    },
    "/api/admin/quotation/{quotation_id}/payment/s": {
      "get": {
        "operationId": "list_payments_api_admin_quotation__quotation_id__payment_s_get",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaymentRead"
                  },
                  "title": "Response List Payments Api Admin Quotation  Quotation Id  Payment S Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List payments for a quotation",
        "tags": [
          "admin:payments"
        ]
      }
    },
    "/api/admin/quotation/{quotation_id}/payment/stripe": {
      "post": {
        "description": "Creates a Payment record (PENDING, type=STRIPE).\n\nThe actual Stripe Checkout Session is created by the B2C client\nwhen they select an auction (see ``app/services/auction.py::\nselect_auction``). This endpoint is the admin-side recording of\nthe payment intent; the Stripe-hosted checkout flow lives on the\nB2C frontend (PR5).",
        "operationId": "create_stripe_payment_api_admin_quotation__quotation_id__payment_stripe_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreateStripe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a Stripe payment record for a quotation",
        "tags": [
          "admin:payments"
        ]
      }
    },
...
    "/api/b2c/api/b2c/quotation/{quotation_id}/invoice": {
      "post": {
        "description": "Capture the client's RFC and create a PENDING invoice for a paid quotation.",
        "operationId": "capture_invoice_api_b2c_api_b2c_quotation__quotation_id__invoice_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCaptureBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capture RFC and create invoice (B2C)",
        "tags": [
          "b2c:invoices"
        ]
      }
    },
...
    "/webhooks/payments/stripe": {
      "post": {
        "operationId": "stripe_webhook_webhooks_payments_stripe_post",
        "parameters": [
          {
            "in": "header",
            "name": "Stripe-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stripe-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stripe Webhook",
        "tags": [
          "webhooks"
        ]
      }
    }
  }
}